pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "work-log-synchronizer"
version = "0.1.0"
description = "Synchronize work logs from Clockify to BambooHR timesheet"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [{name = "Work Log Sync"}]
dependencies = [
    "typer[all]~=0.14.0",
    "httpx~=0.27.0",
    "pydantic~=2.10.0",
    "pydantic-settings~=2.5.0",
    "pyyaml~=6.0.2",
    "python-dateutil~=2.8.0",
    "rich~=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest~=8.0.2",
    "pytest-cov~=4.1.0",
    "pytest-asyncio~=0.23.3",
    "ruff~=0.6.9",
    "mypy~=1.13.0",
    "types-PyYAML~=6.0.0",
    "types-python-dateutil~=2.8.0",
]

[project.scripts]
work-log-sync = "work_log_sync.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/work_log_sync"]

[tool.ruff]
line-length = 100
target-version = "py313"
include = ["src/**/*.py", "tests/**/*.py"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["work_log_sync"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/work_log_sync --cov-report=html --cov-report=term-missing"
